Use boolean in cross-origin reportEvent() opt in header.
The opt in header currently expects the string literal "true" to opt
the fenced frame's config into being used for reportEvent() calls
originating from cross-origin subframes. Using string literals are an
outdated syntax. Instead, the current spec prefers using the boolean
values "?0" and "?1". This CL updates the header to expect a boolean.

The feature has not shipped yet, so there is no need for backwards
compatibility for the existing syntax.

Change-Id: I2d2aea9713b98acb6984f7c57971413fe684afb3
Bug: 331844131